+++ window_tagger/zones.json
@@ -1,6 +1,6 @@
 {
   "centered": {
-    "name": "Centered",
+    "name": "centered",
     "x": 630,
     "y": 20,
     "width": 1493,
